refactor: move shared code to AbstractCompressorBlockEntity
